requirements.txt
streamlit
torch
torchvision
matplotlib
numpy<2
pandas
Pillow
